refactor: moved interfaces
